feat: restructure to have pages
